Limit param of RSS plugin is now an Integer
Close #1051


git-svn-id: https://svn.merengueproject.org/trunk@2618 5c12d508-d76f-0410-91c7-e3e856e0eb0f

+++ merengueproj/plugins/rss/config.py
@@ -51,9 +51,9 @@ class PluginConfig(Plugin):
         params.List(name="contenttypes",
                     label=_("Content types that whant to be syndicated"),
                     choices=get_all_children_classes),
-        params.Single(name="limit",
-                      label=_("number of elements to show at the feed"),
-                      default=10),
+        params.Integer(name="limit",
+                       label=_("number of elements to show at the feed"),
+                       default=10),
     ]
 
     @classmethod
